pyproject.toml
[tool.poetry]
name = "retail-chain"
version = "0.1.0"
description = ""
authors = ["siddhant-mi <siddhant.pandey@mindinventory.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.14"
sentence-transformers = "^2.6.1"
scheduler = "^0.8.5"
faiss-cpu = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
